requirements.txt
nltk
rouge
pandas
sentence_transformers
rank-bm25
tenacity
openai
anthropic
Pillow>=10.3
opencv-python>=4.9
anthropic>=0.31.0
openai>=1.40.0